refactor: NewEventModal em arquivo proprio, aceitando clientId
